Using float instead of double precision float for speed.

+++ filter.c
@@ -2,10 +2,10 @@
 #include "filter.h"
 #include "config.h"
 
-static double ProcessLPF(double x, struct filter_t *lpf);
-static double ProcessHPF(double x, struct filter_t *hpf);
+static float ProcessLPF(float x, struct filter_t *lpf);
+static float ProcessHPF(float x, struct filter_t *hpf);
 
-void CreateFilter(struct filter_t *f, double fc, enum filter_type type)
+void CreateFilter(struct filter_t *f, float fc, enum filter_type type)
 {
 	f->yn = 0;
 	f->xn = 0;
@@ -13,7 +13,7 @@ void CreateFilter(struct filter_t *f, double fc, enum filter_type type)
 	f->type = type;
 }
 
-double ProcessFilter(double x, struct filter_t *f)
+float ProcessFilter(float x, struct filter_t *f)
 {
 	switch (f->type)
 	{
@@ -27,18 +27,18 @@ double ProcessFilter(double x, struct filter_t *f)
 	}
 }
 
-static double ProcessLPF(double x, struct filter_t *lpf)
+static float ProcessLPF(float x, struct filter_t *lpf)
 {
-	double y;
+	float y;
 	y = lpf->yn + lpf->k * (lpf->xn - lpf->yn);
 	lpf->xn = x;
 	lpf->yn = y;
 	return y;
 }
 
-static double ProcessHPF(double x, struct filter_t *hpf)
+static float ProcessHPF(float x, struct filter_t *hpf)
 {
-	double y;
+	float y;
 	y = x - hpf->k * hpf->yn;
 	hpf->yn = y;
 	return y;
