+++ notes.c
@@ -1,7 +1,7 @@
 #include <math.h>
 #include "notes.h"
 
-double Note2Freq(uint8_t note)
+float Note2Freq(uint8_t note)
 {
 	return pow(2, (float)(note - 69) / 12) * 440;
 }
